fix: update curl to follow redirections
Due to MaxMind update on the 1st May 2024 :
```txt
Hello,

As of Wednesday, May 1, 2024, all MaxMind database downloads will redirect to R2 presigned URLs. We are making this change in order to increase the security and reliability of our services. 

This is a potentially breaking change. Please ensure your downloading HTTP client is configured to follow redirects and can connect to the following hostname:

 mm-prod-geoip-databases.a2649acb697e2c09b632799562c076f2.r2.cloudflarestorage.com

We recommend confirming the above as early as possible.

You do not need to change the URL you download from. Existing permalinks will continue to work.
```

The file is downloaded after recieving a 302 👌